requirements.txt
 pandas==2.2.3
numpy==2.2.5
matplotlib==3.10.1
seaborn==0.13.2
scikit-learn==1.6.1
xgboost==3.0.0
lightgbm==4.6.0
joblib==1.5.0
streamlit==1.45.0
jupyter==1.1.1
pre-commit==4.2.0
black==25.1.0
flake8==7.2.0
nbqa==1.9.1
plotly>=5.0.0
pyyaml==6.0.2
pytest==8.3.3
pytest-cov==5.0.0
